Fix tree generation bugs and code issues
- Remove unused UV variable v in branch geometry generation
- Fix Space Colonisation trees to respect barkTint parameter
- Fix Space Colonisation trees to respect barkTextureScale parameter
- Remove unused NoV variable in tree.frag and tree_billboard.frag
- Change trunkSegments from float to int (semantic correctness)
- Clamp leafStartLevel when branchLevels is reduced

Note: Shaders must be recompiled after these changes.

+++ src/TreeGenerator.h
@@ -91,7 +91,7 @@ struct SpaceColonisationParams {
     float tropismStrength = 0.1f;       // How much tropism affects growth
 
     // Initial trunk before branching
-    float trunkSegments = 3;            // Number of trunk segments before crown
+    int trunkSegments = 3;              // Number of trunk segments before crown
     float trunkHeight = 2.5f;           // Height of trunk before crown starts (~2.5m)
 
     // Root system
